Skip screenshot checks for strings marked with is_accessibility=true
We require translation screenshots for strings in grd and grdp files.
It's been noted that some of these strings are accessibility labels, and
they aren't displayed in Chrome UI.

This CL introduces a new tag to mark strings that are used as
accessibility labels (is_accessibility=true). If a UI message has this
tag, it'll be excluded from translation screenshot checks in the
presubmit.

Bug: 1094077
Change-Id: I204d068264ff0d0305d1d8afd48de516f9a26ac1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274667
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784078}