cc: Fix retention on YUV planes in PaintImage
When an SkImage is created with YUV planes, the caller is responsible
for lifetime of those planes. This change ensures that texture backed
PaintImages maintain a reference to the YUV planes |sk_planes_| used to
generate an SkImage as well as the |sk_image_| itself.

Bug: 1091347
Change-Id: If253b3f4d047432dc4ab48d99563168ee7d5ea1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264259
Commit-Queue: Sasha McIntosh <sashamcintosh@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784119}