goma_link: add --allowlist option
During testing and troubleshooting, it is often useful to have
goma_link.py and goma_ld.py perform distributed ThinLTO for targets
that are not on the allow list. This change adds an --allowlist option
which, when given, causes the scripts to proceed as if the target
being built is on the allowlist, regardless of whether it actually is.

Bug: 1072888
Change-Id: Ide4f273b022d3451cc2eb5341a3ee5e0be1ff323
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275633
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784209}