[Extensions] Re-enable some tab capture tests.
These tests were disabled for various reasons, but passed 3000
iterations on bots for all platforms.

Note that sub-tests that checked for permissions failures were
removed, since the <all_urls> permission is now needed for this
API, and that was added to the manifest for other tests that
weren't disabled.

Bug: 80212, 117927, 367695
Change-Id: I074f78626cac1fa4c8be6e020ea77b7e8ca3d65c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2267929
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784236}