Abstract away how the toolbar icon container observes changes.
This will allow changing the observation method in the future without
affecting the callers.  This does require changing the pointer type in
PageActionIconParams to a concrete type, but that's OK for now; if we
really need to later, we could introduce a new abstract class that
simply exposes an ObserveButton() method.

Bug: none
Change-Id: I540905bf89d0d05ecb69806495d9a3618b21a19f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274842
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784270}