Set has_committed_any_navigation_ for error pages.
Previously, this was only set from CommitNavigation.  Error pages
go through FailedNavigation instead of CommitNavigation, so they were
never setting this to true.  Looking at the use cases of
has_committed_any_navigation_, it seems that this is a bug, so this CL
fixes it.

This is a prerequisite for https://crrev.com/c/2250945, which will avoid
BrowsingInstance swaps for browser-initiated navigations if
has_committed_any_navigation_ is false, to fix issues with improperly
using a spare process for initial RFHs.

Doing this also makes the test
SitePerProcessBrowserTest.ProcessTransferAfterError fail with
RenderDocument for subframes due to https://crbug.com/1068965, so
that test is disabled in that mode.

Bug: 1094088,1068965
Change-Id: I942fd4046bb1c6941cf5c8aaf1c17b37ed9cdbdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2272029
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784276}