Implement predicate-based decoder selection
This CL revamps the way a VideoDecoder or AudioDecoder is selected to
decode a given VideoDecoderConfig or AudioDecoderConfig, respectively.
The goal is to first filter and rank available decoders based on their
high-level attributes ("IsPlatformDecoder", "SupportsDecryption"),
rather than attempting to initialize them in the order they're
submitted. The result is that an initial decoder is selected more
efficiently, and the decoding system can gracefully transition between
platform and non-platform decoders if the config changes mid-stream.

The full design doc is available at:
go/predicate-based-decoder-selection

Bug: 684792
Change-Id: I8660e7315881b2bea44bf52afb4505f979a43728
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247331
Commit-Queue: Will Cassella <cassew@google.com>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784292}