Implement HTTPSOCSPTests as browsertests
Right now the most complete set of OCSP tests is implemented in
url_request_unittest.cc using HTTPSOCSPTest. This CL creates a parallel
set of browsertests, so we can test that OCSP works with the
CertVerifierService as well.

The new tests are in ocsp_browsertest.cc, and existing OCSP
browsertests were moved from ssl_browsertest.cc into
ocsp_browsertest.cc to reduce the size of the ssl_browsertest.cc
file.

The new tests are nearly identical, structurally, to the original
tests, but SystemSupportsOCSP() is no longer necessary as it
was really checking support for EV certificates, and the checks
for EV certificates are already guarded by
SystemUsesChromiumEVMetadata().

Bug: 1015134, 1022561
Change-Id: I32e9355d276f75de3432c09f8e7309db1525f729
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274581
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784338}