[css-flex] Remove death paths when evaluating percentage resolution
Both CrossSizeIsDefiniteForPercentageResolution and
MainSizeIsDefiniteForPercentageResolution() are only called from
UseOverrideLogicalHeightForPerentageResolution(). The former is called
whenever MainAxisIsInlineAxis(child) is true and the latter when it's
false. However CrossSizeIsDefiniteForPercentageResolution() has a path
for MainAxisIsInlineAxis(child)==false which is impossible to reach. Thesame happens to MainSizeIsDefiniteForPercentageResolution() which has a path for MainAxisIsInlineAxis(child)==true which is also impossible to
reach.

Remove both death paths and replace them by assertions.

Change-Id: I7c97ee46e992585ff2ff9b95e23341cc10b70eed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275446
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#784369}