[Autofill Assistant] Improve overlay behavior in browse mode
Currently, the overlay is depending on the current state. In the case
of browse mode, the state is set to BROWSE by the prompt action and
reset to RUNNING at the end of it. This causes the overlay to flicker
between prompt actions when multiple are present.

To prevent this we introduce a new action to manually configure the
overlay state from the backend. When the the new state is set to HIDDEN
the overlay will never be shown. This overrides the current logic that
decides the overlay state based on the current AutofillAssistantState.


Bug: b/157457059, b/156232664
Change-Id: I38a720ec6a4a6d73ef4bf31f3f5286188942464e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252045
Reviewed-by: Marian Fechete <marianfe@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Luca Hunkeler <hluca@google.com>
Cr-Commit-Position: refs/heads/master@{#784380}