mac: Divide the entitlements into Chrome-specific and Chromium files.
This let the signing scripts work out-of-the-box with Chromium builds.
The existing app-entitlements.plist file is split into two: in that file
are all the entitlements that work with any signed build of the product,
but the entitlements that are bound to the Google Chrome signing
identity are moved into a new app-entitlements-chrome.plist.

Bug: 1021255
Change-Id: I2aa04d5da1edcaaeeed50f3b70a86400e3aac516
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276523
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784391}