Fix browser view hittest calculation
The tabstrip background calculation currently assumes that its origin
is relative to the browser view. This is usually true, but its actual
parent view is the top container, which can have a different origin in
some cases.

This change explicitly converts the origin to the browser view.

Bug: 968415
Change-Id: Ice3243c275ffce2f2fcc043df6b142392f02db2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276520
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784394}