[css-pseudo] Use list item computed style when resolving marker margins
The list-style-type and list-style-image properties can affect the kind
of list marker that we have, including its margin. However, what matters
are the values on the originating list item, not on the ::marker itself.

Right now the difference doesn't matter, since ::marker doesn't accept
these properties, and thus must be inherited. But in the future we might
remove the restrictions and accept arbitrary properties in ::marker.

This patch uses the proper ComputedStyle when resolving margins.

Bug: 457718
Change-Id: I41aafb4b0b5778f637af6d74c0f50146a3698be8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2263834
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#784433}