Fix notification title for zero days remaining
As we use std::lround(...) for calculating the number of remaining days
from time remaining to reach the deadline, it can result in return
value of zero if we are very close to the deadline. In this case, we
need to use the last day notification strings (days_remaining = 1).
The plural string IDS_UPDATE_REQUIRED_NETWORK_LIMITATION_TITLE_DAYS
only has two options i.e. 1 and Other.

Bug: 1048607
Change-Id: I60de9aac659503557ccc4ead60a434ae6d5184d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270186
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Cr-Commit-Position: refs/heads/master@{#784443}