[omnibox] Remove references to "master".
Per https://tools.ietf.org/id/draft-knodel-terminology-00.html, we
should not have any uses of this term in our code. There are clearer and
more inclusive ways of describing this concept. In this case, the
geolocation header code is particularly egregious since, aside from the
use of oppressive language, it doesn't even match the terminology in the
histogram enum it corresponds with*, which inhibits readability.

* https://source.chromium.org/chromium/chromium/src/+/master:tools/metrics/histograms/enums.xml;l=32162-32170;drc=1ed3a9904c7e5325fe034dba8817a99c343bd484

Change-Id: I83c38d82d7808964d10505b448af247499980f87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2269406
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Commit-Queue: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784463}