Make comment in MatchesAVIFSignature more accurate
Make the comment in the MatchesAVIFSignature() method more accurate.
Correct the brand that indicates AVIF ('avif' or 'avis', not 'av01'.)
List the fields in the same order as they appear in the byte stream.
Change "version" to "minor version".

Note: We could get rid of the hardcoded 144-byte maximum length by
reading the 'size' field of the 'ftyp' box as in
https://crrev.com/c/2277182. I can attempt that in a separate CL.

Bug: 960620
Change-Id: Ic99a4045282e59ad0009d0ce6647f3106fa185a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278163
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784547}