Use inode/directory mime type for directories
We are currently using application/octet-stream for directories
which is incorrect.  This changes to use inode/directory mime type
as per XDG spec:
https://specifications.freedesktop.org/shared-mime-info-spec/shared-mime-info-spec-latest.html#idm45070737701600.

Some apps like vscode already register to handle this mime type so
that they can be used to open a directory.  This change will
trigger FilesApp to now allow 'Open with' on directories and show
vscode and any other app registered.

Bug: 1097715
Change-Id: Ibafbc16a69d93133b02c0fa92b5e28dd3281034c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2272418
Reviewed-by: Ben Wells <benwells@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784621}