Add the RawIconPngDataToImageSkia interface for the standard icons.
This is the preparing for CL:2251600.

The new RawIconPngDataToImageSkia interface is used to read the png data
from the arc mojom structure RawIconPngData, and generate the standard
icon:
1. For non-adaptive icons, add the white background and apply the mask.
2. For adaptive icons, read the foreground and background png data, and
composite them, then apply the mask.

The new interface will be used for the arc shortcut, search result, and
windows icons, etc.

BUG=1083331

Change-Id: I592b428bd05397051178e0f00717c0c1465cadb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277368
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784634}