Modify the AppWindowBase interface to use ImageSkia for icons.
This is the preparing for CL:2251600. ARC mojom interface
OnTaskDescriptionUpdated will be modified, and use AppService interface
to load the ARC adaptive icon, return the ImageSkia image, so modify the
related interfaces to use the ImageSkia class, including:
AppWindowBase::SetDescription
ArcAppWindow::SetDescription
ArcAppWindowInfo::SetDescription

Modify the browser test to RunUntilIdle because background thread is
used to decompress the image.

arc_app_window_launcher_controller.cc is modified to pass the build, and
it will be deleted soon when removing the AppServiceInstanceRegistry.

BUG=1083331

Change-Id: I4e232c64811485c963b6e5584a55d4fa3e07ef64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275166
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Long Cheng <lgcheng@google.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784657}