[FragmentItem] Clear inline fragment index for positioned objects
When once laid out floating object is changed to a positioned
object, it may have |FirstInlineFragmentItemIndex| set, but
positioned objects should not have the index because they do
not produce |NGFragmentIem|. This patch makes sure the index
is cleared for positioned objects.

Also this patch turns DCHECK in |NGInlineCursor::MoveTo| to
NOTREACHED + return to avoid such case could result in crash.

This is a regression in r784297 <crrev.com/c/2275373>, which
started to index floating objects.

Bug: 1101277
Change-Id: I21a75440c3aa77a8d66b1e002d05ff2b252e6b94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275820
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784685}