arc-a11y: Update window id mapping for ARC-R
From ARC R, task id is not sent from mojo, and window id is used to map
the sent event to task id.
CL:1982507 introduced the basic implementation for it, but it doesn't
work well when a mojo event arrives before the exo property is updated.

This CL changes:
- Keep mapping from window id to task id in a11y layer so that changing
  window id back to what was already used before won't drop any events.
- When exo window id property is updated, if the window id is new to
  a11y, request Android to send the whole a11y tree so that the tree
  status is kept in sync.

AX-Relnotes: n/a. This is a preparation for ARC-R.
Bug: b:157212802
Test: ArcAccessibilityHelperBridgeTest, ArcAccessibilityHelperBridgeBrowserTest
Change-Id: Icc68d930dcbe7eb5d81674daa01fa99beea474e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274487
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: Sara Kato <sarakato@chromium.org>
Reviewed-by: Daichi Hirono <hirono@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784686}