WebShare: No longer experimental
WebShare was being exposed in IDL on all platforms for users with
experimental web platform features enabled, causing a crash if
navigator.share() is called with valid arguments, as the mojo
service is not present.

Apart from Android, we now only expose the feature for tests.

Bug: 736172,1099095
Change-Id: Ide6ba0e9620a1ee32163f4a4227a15372d1d1e0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275564
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Eric Willigers <ericwilligers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784699}