[IntentHandling] Add activity name for intent filter from ARC.
This CL receives the activity name for intent filters from ARC. At the
moment, only share target filters will have activity name. This is
needed to show on the sharesheet to show the actual activity it is about
to perform. Also it is useful to show the activity only once for
multiple intent filter matches.

Test: Manually print out and see the activity name passed over.
BUG=1092784

Change-Id: If64d0c5acbf48c0c205eeeb8b9f3c09cac563964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2272619
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Reviewed-by: David Jacobo <djacobo@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784706}