[IntentHandling] Create intent with filesystem URLs and mime types.
This CL adds the support to create intent with filesystem URLs and mime
types. To do this, a filed called file_urls is added in the intent
struct to hold the data.

When sharing multiple files from Files app, a general intent from the
MIME types of all the files is created, and if the MIME types are
different, this creates a wildcard intent matching MIME type */*.
However, this leads to a bug where the wildcard matches intent fields
of the form supertype/* in Android - e.g. the Android YouTube app
(which accepts video/*) will accept shares of a multiply-selected
different MIME types from Files app, even if those files are not video
files.

Therefore, in my code, at the moment there is no need to provide
wildcard matching for intent, so this CL updates the matching rule to
only have wildcard for intent filter, i.e. if intent is */*, it will
only match with */* intent filter to make it correct (but not the same
with Android implementation). This is a simpler change compare with
carrying multiple mime types in the intent, unless we want to handle
file that contain no mime type but want to be handled by all filters
that can handle files (even if they have specific types).

BUG=1092784

Change-Id: Iec607bd1c117040b48a326096d82173d426d1598
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275100
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784707}