Make sure SetIdleDelay is called after a service worker is started
In renderer side, SetIdleDelay is expected to be called after
ServiceWorkerEventQueue::Start() is called. However, it can be called
during startup when ServiceWorkerTerminationOnNoControllee is enabled.

This CL fixes the problem by checking the running status when
AddControllee() and RemoveControllee(), and sending the right value
after starting a service worker.

Bug: 1099744
Change-Id: If932c1eb73ea53ad18162324f32c23f17c46d49b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275982
Commit-Queue: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784719}