[menus] Position clipped sub menu on widest side
When sub menus can't fit between the right hand side of the parent menu
and the limit of the window they are placed on the left. In the case
where we have a very skinny window there could be less space to the
left of the parent menu and so the result is more clipping of the
sub menu when placed on the left.

Change the logic that detects sub menu clip case and position it on the
side that has more space available.

Bug: 1100222
Change-Id: If93561fee074173c995b30fad32c15f48e2f22ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278908
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784727}