Files app: Fix list scroll for change default app dialog
Add "pointer-events: none;" in the pseudo-element, to fix the clickable
area for the last item in the list that. The pseudo-element covers the
last item, so we avoid taking the click in the pseudo-element.

Add scroll event listener to detect when the scroll has reached the
bottom of the list and remove the CSS class "bottom-shadow" which
controls when the gradient shadow should be displayed in the bottom of
the list.

Tested manually, integration tests coming in the dependent CL.

Bug: 1100164
Change-Id: Id27dca900f850fec088a5567182bcb4276f7264f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278745
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Auto-Submit: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784729}