[Android translate infobar] Check items for never translate lang/site
On Android, the translate infobar's overflow menu has the "always
translate language" item checked if the corresponding option is set.
However, it does not do so for the "never translate language (site)"
items. It is possible for the user to get in a state where they see
the translate infobar with one of these options set: they can invoke
translate on a site that falls in one of these buckets manually via
the Chrome overflow menu.

I discovered this while implementing the translate infobar in WebLayer;
there the bug was more apparent because the infobar stays present after
the user clicks on one of these options due to the lack of snackbars.
I'm porting the fix to Chrome in advance of componentizing //chrome's
class for sharing with WebLayer.

Bug: 1101240
Change-Id: Ibe12e3a7a77a555c93e068b8181114e3945b3e8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276287
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Megan Jablonski <megjablon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784733}