Avoid passing CHROME_CONNECTED cookie.
This CL avoids sending the CHROME_CONNECTED cookie in requests initiated
from the browser. When this cookie is included, the server blocks the
logout requests and return 204 (this is part of the protocol between
Chrome and Gaia to manage Google authentication cookies).

Note: The tests in gaia_auth_fetcher_ios_ns_url_session_bridge_unittest.mm
are very flaky and several attempts to re-enable them failed. Unit test
for this bug will be added once the rest of the tests are stable.

Bug: 1101704

Change-Id: Ica5eaaea04fe5e3c175737fd5433e1a227988807
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276290
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784756}