Use |registered_origins_| from SWStorage for GetInstalledRegistrationOrigins
The method GetInstalledRegistrationOrigins() currently uses
|registered_origins_| which is initialized on the UI thread in
ServiceWorkerContextWrapper, and thus should only be accessed from the UI
thread. This means that a task should also be posted to the UI thread.

However, the ServiceWorkerStorage also has |registered_origins_| which
can be used by calling ServiceWorkerStorage::GetRegisteredOrigins() from
ServiceWorkerContextWrapper::GetInstalledRegisteredOrigins() on the Core
thread. Hence, this was deemed to a better solution to the problem as
this avoids thread hopping. This CL implements this change.

Bug: 807440
Change-Id: Ib6d9b2adee7daaf7b3560774caf17295e5d1a707
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277689
Commit-Queue: Nidhi Jaju <nidhijaju@google.com>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784766}