Clean up unused profile menu strings
No behavioral changes.

Most callers to the former GetMessagesForAvatarSyncError(), now renamed,
were not interested in the string IDs returned by the function. The only
exception is ProfileMenuView which not only ignored one of the two IDs,
but can also infer the second (the button's string) from the returned
enum.

This avoids API redundancy and aligns better with existing assumptions
in ProfileMenuView, such as OnSyncErrorButtonClicked() using the enum
to determine what the button's action should be.

The side effect is that some strings can be cleaned up.

Change-Id: I2c3b47c47e06e0a4a79d3efa3d3ecda498ef489c
Bug: None
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278249
Commit-Queue: David Roger <droger@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784787}