ui/display: make kUseHDRTransferFunction a base::Feature and disable it
This CL changes the switch kEnableUseHDRTransferFunction to a
base::FeatureFlag kUseHDRTransferFunction which is more adequate
to after-launch, where users will like to see the hdr-ness state
as "disabled" "enabled" or "default", as opposed to having a naked
flag where the correspondence chrome:flags<->feature status is
non-intuitive.

This CL also disables kUseHDRTransferFunction with the idea of fixing
the current quirks (e.g. b/158777601, b/158781255, b/159224397) and
enabling it very soon.

This CL also takes the chance to remove the unused entry
display::features::kUseMonitorColorSpace.

TBR=achuith@chromium.org for chrome_restart_request.cc

Bug: 958166, b:159919063
Change-Id: I10e48aee423924cfa7eea7c72d5455f1ffe72648
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276516
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784790}