[WebLayer] Test behavior of intent launching in onLoad() properly
When I changed embedder-initiated navigations to be of type link rather
than user typing, I also changed the ExternalNavigationTest case that
tested handling of intents initiated via onLoad() in a page that is
reached via an embedder-initiated navigation to expect an intent being
launched rather than blocked. This makes sense since it is only for
navigations initiated by user typing that such intents would be blocked.
I should have also changed this test to expect that the visible URL
would revert to that of the original URL, since in this case WebLayer
eliminates the navigation entry that corresponds to the intent being
launched. I neglected to do this and the test itself didn't complain
because it wasn't set up to wait long enough for the navigation to the
page launching an intent to fully complete, including the elimination
of the corresponding navigation entry.

There is no production problem here; this is simply a matter of test
precision that nator@ uncovered.

Change-Id: If1443d202cffd0863af2600094bd46a2663bf2b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279200
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784794}