[COOP] Turn the flag on by default.
COOP was approved to launch in M82. The current development version is
M86.

This patch turns the flag on by default.
This is still disabled on Android WebView.

Since the parser is already behind a flag, there is no need to check for
the flag again and again everywhere in the code. Only some DCHECK are
necessary.

Bug: chromium:922191
Change-Id: I5f43652e38d4df5e003eb36b82eccc6f0a952cd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274882
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Pâris Meuleman <pmeuleman@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784831}