[iOS] Change the NTP animation to use an animator
This CL updates how the NTP animations to shift tiles up (i.e. when the
omnibox is focused), using an animator instead of the regular UIView
-animate:completion:.
It allows to stop the animation in the middle without having the
animated properties in an undefined state.

Bug: 1101331
Change-Id: Iadff66283784c7a339770b58270099678cf2616c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278253
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784857}