[COOP] access reporting [2/N] Browser -> renderer(s) IPC.
All of this is put behind a flag disabled by default.

This is mostly based on the initial prototype:
https://chromium-review.googlesource.com/c/chromium/src/+/2223934/24

Add the IPC InstallCoopAccessMonitor to the LocalMainFrame interface.
This is called by the browser process whenever an access in between two
windows would be blocked if COOP was enforced.

This is mostly based on the initial prototype:
https://chromium-review.googlesource.com/c/chromium/src/+/2223934/24

COOP access reporting:
[1/N] https://chromium-review.googlesource.com/c/chromium/src/+/2264294
[2/N] this patch.

TBR=mkwst@chromium.org

Bug: chromium:1090273
Change-Id: I366ef78ec5ac59ca9475871928afd05d2e02fddd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270185
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Auto-Submit: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784871}