Use page URL to persist Android dependency graph filter
A new UI control is added which will generate a URL based on the current
filters applied to the graph. The URL uses shortened names, and is
parsed on page load to generate the filter for the graph's nodes. If
there is no filter in the URL, some default testing values are used
instead.

Bug: 1093962
Change-Id: I8149c6fd41c2bf84a21f969103146bbd9e2f80bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276675
Commit-Queue: James Long <yjlong@google.com>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784886}