app_list: Preserve selected search result when results change
Instead of always selecting the first result,
ResultSelectionController::ResetSelection(), which is called whenever
the list of search results change, should try to keep the selection on
the currently selected result ID if the selection was result of a user
action (i.e. if currently selected result is not the default result).
The selection is reset to the first available result only if the user
selected result is no longer in the updated result list.

BUG=1099506

Change-Id: I5d15276b00b080726a3fdef36fbbb8eaa447a842
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204968
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784946}