Files app: File list header and icons spacing
Change the file list header spacing to have the same numbers that the
spec: Remove the 8px padding from the non-NG style and make the margin
explicitly 16px.

Change the "detail-icon" width to 40px and fix its margins to 8px like
in the spec.

Change the "check-mark" icon width to 40px This interfered with the Grid
View check-mark, so forced the grid view to be 32px wide.

Change the thumbnail icon margin to 8px. Matching the spec and the
"detail-icon".

Bug: 1065756
Change-Id: Id4385544dcbcaf4c027243792c6ad1277e35b740
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275989
Auto-Submit: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784947}