Move HttpRequestHeaders logic to signin::RequestAdapter
signin::RequestAdapter no longer needs to handle modifying headers on
net::URLRequest, so it should be safe to remove this functionality. In
addition, the header modification logic was being duplicated in both
ThrottleRequestAdapter and ProxyRequestAdapter (and soon to be the
WebLayer version of the throttle adapter). This shares the logic in the
base signin::RequestAdapter.

This also moves the kChromeManageAccountsHeader constant into
//components so it can be shared between Chrome and WebLayer.

Bug: 1054160
Change-Id: Ib015d7bd453ee102ce7a1e3aea2d994f92741b77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276708
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784949}