[FragmentItem] Reduce unnecessary LayoutObjectWillBeDestroyed
This patch removes the call to |NGFragmentItems::
LayoutObjectWillBeDestroyed| in |LayoutText::
RemoveAndDestroyTextBoxes| when |DocumentBeingDestroyed|.

This was added in r756210 <crrev.com/c/2117558>, probably
while trying to pass tests, but there's no good explanation
to need it while destroying the document, and tests pass.

Remove this to see if this has effects on UMA or crashes.

Bug: 982194
Change-Id: I13e8bd7a8d556fafe96b17944ad187354073cd79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276369
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784963}