Vulkan: tracking damage area in SkiaOutputDeviceVulkan
Since max_frames_pending is 1 for SkiaOutputDeviceVulkan, so chrome will
not start a new frame until SwapBuffersAck is received, so we can use
tracking damage area in SkiaOutputDeviceVulkan, and pass the damage area
to SkiaOutputSurfaceImpl via SwapBuffers completion callback. With it,
VulkanSwapChain doesn't need to make sure the acquired images are in
the present order. So remove related code as well.

Change-Id: Iee95d3e43fcc5c6f79c443f9f07b132c29aa8a53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277976
Reviewed-by: Peng Huang <penghuang@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Peng Huang <penghuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784980}