[menus] Allow sub menus to slightly overlap
When the files app window is shrunk horizontally to its minimum width,
any sub menu hanging off the context menu can get clipped.

Changes the positioning logic for the sub menus to allow them to overlap
their parent context menu by up to 16px if they would be clipped by the
files app window.

Note, this is a mitigation for an edge case - this can only happen when
the splitter is pushed as far to the left as possible and the window
narrowed to as small as possible too so unlikely to happen very often
with normal usage.

Bug: 1100222
Change-Id: I547e93b7bc7f7230d33d8c63d5a4d223959f4732
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279293
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784994}