Online re-authentication on the lock screen for SAML users.
In order to sync SAML user's password between multiple CrOS devices we
are introducing a new entity that can send the user via online signin
flow on the lock screen. If the password was changed from other user's
device (e.g. via in-session password change feature) or directly with
the IdP the re-authenication will effectively sync the device with the
IdP.

In the first phase (this CL) online signin will be triggered by the
existing SAMLOfflineSigninTimeLimit policy. In the next step we add
new password sync token checking mechanism (go/device-token-api-dd)
that allows the device to detect password change event and execute
the sync when needed.

Bug: 1090341
Change-Id: I5c707e8365dd1c34acb1e22754691a8f9e720fdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231220
Commit-Queue: Maciek Slusarczyk <mslus@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785014}