[Passwords] Instantiate proper form fetcher in CredentialManagerImpl
Before this CL:
CredentialManagerImpl always instantiated single store FormFetcher
(FormFetcherImpl). Therefore, the "Save bubble" was shown even if
the credentials exist in the account store since only the profile
store was checked for the existence of the credentials.

After this CL:
CredentialManagerImpl uses the factory method
FormFetcherImpl::CreateFormFetcherImpl() that instantiates the
multi-store FormFetcher version if required. This quarantees that the
"Save bubble" will be shown if the credentials don't exist in any
of the stores.


Bug: 1093286
Change-Id: I37840965057ca9aea33bccdbf0a162d6fa9c040f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279871
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785015}