[iOS CBCM] Move browser report generation logic to delegate
In preparation for moving browser_report_generator to components, this
change moves all platform-specific logic to a new delegate class that
will stay in chrome/browser. Currently, the parent object,
BrowserGenerator, is responsible for passing the delegate to
BrowserReportGenerator, but this will be moved to a factory class in a
subsequent change.

Bug: 1091895
Change-Id: Iaa05963737764b169d65ae76cc31e2e318c811df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2280208
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785101}