plugin_vm: implement relaunch flow
This implements relaunch flow for Parallels VMs when changing microphone
and camera permissions, and updates messaging accordingly.

Note that changes in PluginVmManager are kept to bare minimum to
minimize the risk of disturbing existing flows to ease the merging task.
They will need to be enhanced to handle suspending the VM when it
already is transitioning its state, and provide better feedback to the
user on failures.

Most of the work here was done by Daniel Ng <danielng@google.com>

Bug: 1071872
Change-Id: Iebf694706d0c7d3d63fc995ec63180ba85d970a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278320
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Daniel Ng <danielng@google.com>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785190}